requirements.txt
pandas==2.1.4
pyarrow>=18.0.0
google-cloud-bigquery==3.40.1
google-cloud-storage>=2.14.0
requests>=2.28.0